+++ libs/pen-parser/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @protolabsai/pen-parser
 
+## 0.53.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @protolabsai/types@0.59.0
+
+## 0.53.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @protolabsai/types@0.58.0
+
 ## 0.53.4
 
 ### Patch Changes